% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biod}
\alias{biod}
\title{Test community matrix and phylogeny}
\format{
A list with two elements:
\describe{
\item{phy}{Phylogeny with 31 tips}
\item{comm}{Community matrix with 513 sites and 31 species.
Data are counts, i.e., the number of times
a species was observed in a grid cell.}
}
}
\source{
\href{https://github.com/shawnlaffan/biodiverse}{Biodiverse}
}
\usage{
biod
}
\description{
Test data from \href{https://github.com/shawnlaffan/biodiverse}{Biodiverse}
}
\examples{
# Example phylogeny
library(ape)
biod$phy
# Example community
biod$comm[1:5,1:5]
}
\keyword{datasets}
