% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biodiv_random.R
\name{calc_biodiv_random}
\alias{calc_biodiv_random}
\title{Calculate diversity metrics for a single random community}
\usage{
calc_biodiv_random(
  comm,
  phy,
  phy_alt,
  null_model = c("frequency", "richness", "independentswap", "trialswap"),
  n_iterations = 1000,
  metrics = c("pd", "rpd", "pe", "rpe")
)
}
\arguments{
\item{comm}{Dataframe or matrix; input community matrix with communities (sites) as rows
and species as columns, including row names and column names.}

\item{phy}{Input phylogeny with total branch length scaled to 1}

\item{phy_alt}{Alternative phylogeny where all branches are of equal length, scaled to 1}

\item{n_iterations}{Number of iterations to use when shuffling random community}

\item{metrics}{Character vector; names of metrics to calculate}
}
\value{
List of vectors. Each vector is a biodiversity metric measured on the
random community, in the same order as the rows in the input community. Names
of the list correspond to \code{metrics}.
}
\description{
For description of metrics available, see \code{\link{cpr_rand_test}()}
}
\examples{
\dontrun{
library(picante)
data(phylocom)
phy <- phylocom$phy
comm <- phylocom$sample
subsetted_data <- picante::match.phylo.comm(phy = phy, comm = comm)
phy <- subsetted_data[["phy"]]
comm <- subsetted_data[["comm"]]
phy_alt <- phy
phy_alt$edge.length <- rep(x = 1, times = length(phy_alt$edge.length))
phy_alt$edge.length <- phy_alt$edge.length / sum(phy_alt$edge.length)
phy$edge.length <- phy$edge.length / sum(phy$edge.length)
calc_biodiv_random(comm, phy, phy_alt, "independentswap", metrics = c("pd", "pe", "pd_alt"))
}

}
\keyword{internal}
