% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canape.R
\name{canape}
\alias{canape}
\title{Run randomization analysis for a set of biodiversity metrics}
\usage{
canape(
  comm,
  phy = NULL,
  null_model = "independentswap",
  n_reps = 100,
  n_iterations = 10000,
  metrics = c("pd", "rpd", "pe", "rpe")
)
}
\arguments{
\item{comm}{Input community matrix in data.frame format (communities as rows,
species as columns, with row names and column names)}

\item{phy}{Input phylogeny}

\item{null_model}{Name of null model to use. Must choose from 'frequency', 'richness',
'independentswap', or 'trialswap' (see picante::randomizeMatrix).}

\item{n_reps}{Number of random communities to replicate}

\item{n_iterations}{Number of iterations to use when swapping occurrences to
generate each random community}

\item{metrics}{Character vector; names of metrics to calculate}
}
\value{
Tibble. For each of the biodiversity metrics, the observed value (_obs),
mean of the random values (_rand_mean), SD of the random values (_rand_sd),
rank of the observed value vs. the random values (_obs_rank), standard effect size
(_obs_z), and p-value (_obs_p) are given.
}
\description{
The biodiversity metrics analyzed include:
\itemize{
\item pd: Phylogenetic diversity (Faith 1992 https://doi.org/10.1016/0006-3207(92)91201-3)
\item rpd: Relative phylogenetic diversity (Mishler 2014 https://doi.org/10.1038/ncomms5473)
\item pe: Phylogenetic endemism (Rosauer 2009 https://doi.org/10.1111/j.1365-294x.2009.04311.x)\
\item rpe: Relative phylogenetic endemism (Mishler 2014 https://doi.org/10.1038/ncomms5473)
}
}
\details{
The default method for generating random communities is the independent swap
method of Gotelli (2000), which randomizes the community matrix while maintaining
species occurrence frequency and sample species richness.
}
\examples{
library(picante)
data(phylocom)
phy <- phylocom$phy
comm <- phylocom$sample
canape(phylocom$sample, phylocom$phy)
}
