% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaper-package.R
\docType{package}
\name{canaper-package}
\alias{canaper}
\alias{canaper-package}
\title{canaper: Categorical Analysis of Neo- And Paleo-Endemism}
\description{
Provides functions to conduct categorical analysis of neo- and paleo-endemism (CANAPE).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joelnitta/canaper}
  \item \url{https://joelnitta.github.io/canaper/}
  \item Report bugs at \url{https://github.com/joelnitta/canaper/issues}
}

}
\author{
\strong{Maintainer}: Joel H. Nitta \email{joelnitta@gmail.com} (\href{https://orcid.org/0000-0003-4719-7472}{ORCID})

Other contributors:
\itemize{
  \item Shawn W. Laffan [contributor, data contributor]
  \item Brent D. Mishler [contributor, data contributor]
  \item Wataru Iwasaki (\href{https://orcid.org/0000-0002-9169-9245}{ORCID}) [contributor]
  \item Klaus Schliep (\href{https://orcid.org/0000-0003-2941-0161}{ORCID}) [reviewer]
}

}
\keyword{internal}
