% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canape.R
\name{categorize_endemism}
\alias{categorize_endemism}
\title{Categorize phylogenetic endemism}
\usage{
categorize_endemism(df)
}
\arguments{
\item{df}{Input data frame. Must have p-values for pe, pe_alt, and rpe.}
}
\value{
Dataframe with areas of endemism categorized.
}
\description{
see:
http://biodiverse-analysis-software.blogspot.com/2014/11/canape-categorical-analysis-of-palaeo.html
}
\details{
(here, PE_orig = pe_obs_p, PE_alt = pe_alt_obs_p, and RPE = rpe_obs_p)
(don't consider 'super endemism' as it doesn't add much meaning)
\enumerate{
\item If either PE_orig or PE_alt are significantly high then we look for palaeo or neo endemism
a)    If RPE is significantly high then we have palaeo-endemism
(PE_orig is consistently higher than PE_alt across the random realisations)
b)    Else if RPE is significantly low then we have neo-endemism
(PE_orig is consistently lower than PE_alt across the random realisations)
c)    Else we have mixed age endemism in which case
i)    If both PE_orig and PE_alt are highly significant (p<0.01) then we
have super endemism (high in both palaeo and neo)
ii)   Else we have mixed (some mixture of palaeo, neo and non endemic)
\item Else if neither PE_orig or PE_alt are significantly high then we have a non-endemic cell
}
}
