% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_higher}
\alias{count_higher}
\title{Count number of times one number is higher than others}
\usage{
count_higher(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{Number to count}

\item{y}{Vector of numbers to compare}

\item{na.rm}{= Logical; should NA values in the comparison
vector be removed before making comparison?}
}
\value{
Number of times x is higher than y
}
\description{
Count number of times one number is higher than others
}
\examples{
{
count_higher(4, 1:10)
count_higher(4, c(1:10, NaN))
}
}
\keyword{internal}
