% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_lower}
\alias{count_lower}
\title{Count number of times one number is lower than others}
\usage{
count_lower(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{Number to count}

\item{y}{Vector of numbers to compare}

\item{na.rm}{= Logical; should NA values in the comparison
vector be removed before making comparison?}
}
\value{
Number of times x is lower than y
}
\description{
Count number of times one number is lower than others
}
\examples{
\dontrun{
count_lower(4, 1:10)
count_lower(NaN, 1:10)
}
}
\keyword{internal}
