% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr_iter_test.R
\name{cpr_iter_test}
\alias{cpr_iter_test}
\title{Test the number of iterations needed to generate a random community that is
sufficiently different from the original community.}
\usage{
cpr_iter_test(
  comm,
  null_model = "curveball",
  binary = NULL,
  thin = 1,
  seed = NULL,
  n_iter_test,
  time_digits = 5,
  time_units = "secs"
)
}
\arguments{
\item{comm}{Dataframe or matrix; input community data with
sites (communities) as rows and species as columns. Values of each cell are
the presence/absence (0 or 1) or number of individuals (abundance) of each
species in each site.}

\item{null_model}{Character vector of length 1 or object of class \code{commsim};
either the name of the model to use for generating random communities (null
model), or a custom null model. For full list of available predefined null
models, see the help file of \code{\link[vegan:commsim]{vegan::commsim()}}, or run
\code{\link[vegan:commsim]{vegan::make.commsim()}}. An object of class \code{commsim} can be generated with
\code{\link[vegan:commsim]{vegan::commsim()}}.}

\item{binary}{Logical vector of length 1; is null model a binary model
(produces random community data matrix with 1s and 0s) or not?
If NULL (default), will attempt to determine automatically from name of
null model (\code{null_model}).}

\item{thin}{Numeric vector of length 1; thinning parameter used by some
null models in \code{vegan} (e.g., \code{quasiswap}); ignored for other models.}

\item{seed}{Integer vector of length 1 or NULL; random seed that will be used
in a call to \code{set.seed()} before randomizing the matrix. Default (\code{NULL})
will not change the random generator state.}

\item{n_iter_test}{Numeric vector; number of iterations to test (see Details
and Examples).}

\item{time_digits}{Numeric vector of length 1; number of digits to include
when recording elapsed time to calculate each random community. Default 5.}

\item{time_units}{Character vector of length 1; units to use when
recording elapsed time to calculate each random community. Must be
a valid value for \code{units} argument of \code{\link[base:difftime]{base::difftime()}}. Default seconds
(`"secs"``).}
}
\value{
Tibble (dataframe) with the following columns:
\itemize{
\item \code{n_iter}: Number of iterations used to generate random community
\item \code{dist}: Euclidean distance between original community and
random community
\item \code{time}: Amount of time elapsed when generating random community
}
}
\description{
A single random community will be created for each value in \code{n_iter_test},
using that value for the \code{n_iterations} argument of \code{\link[=cpr_rand_comm]{cpr_rand_comm()}}. The
Euclidean distance is then calculated between the original
community and each random community.
}
\details{
The user should verify that distances increase with increasing
iterations until an approximate maximum is reached. Of course, this only
makes sense for randomization algorithms that use iterations.

The number of iterations to test (\code{n_iter_test}) is left to the user. A good
rule of thumb is to test successive powers of 10. In this case, no more
than four or five values is usually needed before distances stabilize (see
Examples). Generally, fewer iterations are needed for smaller datasets.
}
\examples{
data(biod_example)
cpr_iter_test(
  comm = biod_example$comm,
  null_model = "swap",
  n_iter_test = c(1, 10, 1000, 10000),
  seed = 123
)

}
