% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpr_signif_cols}
\alias{cpr_signif_cols}
\title{Color palette for plotting results of randomization test}
\format{
An object of class \code{character} of length 5.

An object of class \code{character} of length 5.
}
\usage{
cpr_signif_cols

cpr_signif_cols
}
\description{
Character vector with names corresponding to significance levels
and values corresponding to color codes.

Character vector with names corresponding to endemism types
and values corresponding to color codes.
}
\keyword{datasets}
