% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpr_signif_cols_2}
\alias{cpr_signif_cols_2}
\title{Color palette for plotting results of randomization test, version 2}
\format{
An object of class \code{character} of length 5.
}
\usage{
cpr_signif_cols_2
}
\description{
Character vector with names corresponding to significance levels and values
corresponding to color codes, with "not significant" colored grey.
}
\examples{
cpr_signif_cols_2
\dontrun{
scales::show_col(cpr_signif_cols_2)
}
}
\keyword{datasets}
