% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canape.R
\name{get_ses}
\alias{get_ses}
\title{Extract standard effect size (and other related statistics) for a single
diversity metric given random values and observed values of the metric}
\usage{
get_ses(random_vals, obs_vals, metric)
}
\arguments{
\item{random_vals}{List of list of vectors. Each list of vectors is a biodiversity metric measured on a
random community, in the same order as the rows in the input community.}

\item{obs_vals}{Observed values of the biodiversity metric}

\item{metric}{Name of the metric ("mpd", "mntd", "mpd_morph", "mntd_morph", "pd", "pe", or "rpe")}
}
\value{
Tibble
}
\description{
Extract standard effect size (and other related statistics) for a single
diversity metric given random values and observed values of the metric
}
\examples{
\dontrun{
library(picante)
data(phylocom)
phy <- phylocom$phy
comm <- phylocom$sample
subsetted_data <- picante::match.phylo.comm(phy = phy, comm = comm)
phy <- subsetted_data[["phy"]]
comm <- subsetted_data[["comm"]]
phy_alt <- phy
phy_alt$edge.length <- rep(x = 1, times = length(phy_alt$edge.length))
phy_alt$edge.length <- phy_alt$edge.length / sum(phy_alt$edge.length)
phy$edge.length <- phy$edge.length / sum(phy$edge.length)
	random_vals <-
	purrr::rerun(
		100,
		calc_biodiv_random(comm, phy, phy_alt, "independentswap", 1000, metrics = "pe")
	)
	comm_sparse <- phyloregion::dense2sparse(comm)
	pe_obs <- phyloregion::phylo_endemism(comm_sparse, phy, weighted = TRUE)
	get_ses(random_vals, pe_obs, "pe")
	}
}
\keyword{internal}
