% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mishler_signif_cols}
\alias{mishler_signif_cols}
\title{Original color palette for plotting results of CANAPE}
\format{
An object of class \code{character} of length 5.
}
\usage{
mishler_signif_cols
}
\description{
Character vector with names corresponding to endemism types
and values corresponding to color codes. Original palette used by
Mishler et al. (2014). May not be distinguishable to people with
color-vision deficiency (CVD).
}
\examples{
mishler_signif_cols
scales::show_col(mishler_signif_cols)
}
\references{
Mishler, B., Knerr, N., González-Orozco, C. et al.  (2014)
Phylogenetic measures of biodiversity and neo- and paleo-endemism in
Australian Acacia. Nat Commun, 5: 4473. \doi{10.1038/ncomms5473}
}
\seealso{
Other colors: 
\code{\link{cpr_endem_cols_2}},
\code{\link{cpr_endem_cols}},
\code{\link{cpr_make_pal}()},
\code{\link{cpr_signif_cols_2}},
\code{\link{cpr_signif_cols}},
\code{\link{mishler_endem_cols}}
}
\concept{colors}
\keyword{datasets}
