% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canaper-package.R
\docType{package}
\name{canaper-package}
\alias{canaper}
\alias{canaper-package}
\title{canaper: Categorical Analysis of Neo- And Paleo-Endemism}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to analyze the spatial distribution of biodiversity, in particular categorical analysis of neo- and paleo-endemism (CANAPE) as described in Mishler et al (2014) \doi{10.1038/ncomms5473}. 'canaper' conducts statistical tests to determine the types of endemism that occur in a study area while accounting for the evolutionary relationships of species.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/canaper}
  \item \url{https://docs.ropensci.org/canaper/}
  \item Report bugs at \url{https://github.com/ropensci/canaper/issues}
}

}
\author{
\strong{Maintainer}: Joel H. Nitta \email{joelnitta@gmail.com} (\href{https://orcid.org/0000-0003-4719-7472}{ORCID})

Other contributors:
\itemize{
  \item Shawn W. Laffan [contributor, data contributor]
  \item Brent D. Mishler [contributor, data contributor]
  \item Wataru Iwasaki (\href{https://orcid.org/0000-0002-9169-9245}{ORCID}) [contributor]
  \item Klaus Schliep (\href{https://orcid.org/0000-0003-2941-0161}{ORCID}) [reviewer]
}

}
\keyword{internal}
