% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpr_endem_cols}
\alias{cpr_endem_cols}
\title{Color palette for plotting results of CANAPE}
\format{
An object of class \code{character} of length 5.
}
\usage{
cpr_endem_cols
}
\description{
Character vector with names corresponding to endemism types
and values corresponding to color codes.
}
\examples{
cpr_endem_cols
\dontrun{
scales::show_col(cpr_endem_cols)
}
}
\keyword{datasets}
