% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpr_endem_cols_2}
\alias{cpr_endem_cols_2}
\title{CVD-friendly color palette for plotting results of CANAPE, version 2}
\format{
An object of class \code{character} of length 5.
}
\usage{
cpr_endem_cols_2
}
\description{
Character vector with names corresponding to endemism types
and values corresponding to color codes. Should be distinguishable to
people with color vision deficiency (CVD).
}
\details{
Color scheme:
\itemize{
\item paleo: dark blue
\item neo: red
\item not significant: light grey
\item mixed: green
\item super: yellow
}
}
\examples{
cpr_endem_cols_2
scales::show_col(cpr_endem_cols_2)
}
\seealso{
Other colors: 
\code{\link{cpr_endem_cols_3}},
\code{\link{cpr_endem_cols_4}},
\code{\link{cpr_endem_cols}},
\code{\link{cpr_make_pal}()},
\code{\link{cpr_signif_cols_2}},
\code{\link{cpr_signif_cols}},
\code{\link{mishler_endem_cols}},
\code{\link{mishler_signif_cols}}
}
\concept{colors}
\keyword{datasets}
