% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpr_signif_cols}
\alias{cpr_signif_cols}
\title{CVD-friendly color palette for plotting results of randomization test}
\format{
An object of class \code{character} of length 5.
}
\usage{
cpr_signif_cols
}
\description{
Character vector with names corresponding to significance levels
and values corresponding to color codes.
}
\examples{
cpr_signif_cols
scales::show_col(cpr_signif_cols)
}
\seealso{
Other colors: 
\code{\link{cpr_endem_cols_2}},
\code{\link{cpr_endem_cols_3}},
\code{\link{cpr_endem_cols_4}},
\code{\link{cpr_endem_cols}},
\code{\link{cpr_make_pal}()},
\code{\link{cpr_signif_cols_2}},
\code{\link{mishler_endem_cols}},
\code{\link{mishler_signif_cols}}
}
\concept{colors}
\keyword{datasets}
